requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pathspec==0.6.0
pluggy==0.13.1
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.5
pytest==5.3.2
python-dotenv==0.10.3
regex==2019.12.20
six==1.13.0
SQLAlchemy==1.3.12
toml==0.10.0
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
